scova_backend/requirements.txt
# --- Web framework ---
Flask==3.1.1
flask-cors==6.0.2
gunicorn==23.0.0

# --- Config & HTTP ---
python-dotenv==1.1.0
requests==2.32.3

# --- Database ---
SQLAlchemy==2.0.41
psycopg2-binary>=2.9.10

# --- Supabase storage ---
supabase

# --- Document processing ---
PyMuPDF==1.25.5
python-docx==1.1.2

# --- NLP / Scoring ---
numpy<2.0
Sastrawi==1.0.1

# --- AI / Embeddings ---
google-genai

# --- Testing ---
pytest==8.4.0
